% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisectError.R
\name{bisectError}
\alias{bisectError}
\title{Private function to search for defective wells}
\usage{
bisectError(f, l, min = 1L, max = length(l), prevMin = NA, prevMax = NA)
}
\arguments{
\item{f}{A function that crashes on some wells}

\item{l}{A list of wells}

\item{min}{The smaller index to start with (defaults to 1)}

\item{max}{The larger index to stop with (defaults to list length)}

\item{prevMin, prevMax}{Arguments used to pass previous state in recursion}
}
\value{
A numeric vector of two integers representing the smallest range of
indices necessary to crash the function.
}
\description{
Sometimes defective wells cause some functions to crash.  Given a list
of wells, this function will bisect the list until it converges to the
well(s) that cause a given function to crash.
}
\note{
This function can surely be used in a broader way.  Let me know if
you would like to move it to another package.
}
\author{
Charles Plessy
}
\concept{Private functions}
