% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCR.R
\name{calc_qPCR_mod}
\alias{calc_qPCR_mod}
\title{Compute CT and amplification efficiency from raw data}
\usage{
calc_qPCR_mod(fluo)
}
\arguments{
\item{fluo}{A table providing the raw data loaded by \code{\link{load_qPCR_raw}}.}
}
\value{
A \code{tibble} providing \code{Eff} (efficiency), \code{cpD1} (CT), and \code{cpD2}
another computation of CT) calculated by the \code{\link[qpcR:efficiency]{qpcR::efficiency}} function.
}
\description{
Compute CT and amplification efficiency from raw data
}
\seealso{
Other qPCR functions: 
\code{\link{calc_qPCR_TM}()},
\code{\link{load_qPCR_results}()}
}
\concept{qPCR functions}
