% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo.R
\name{changeWell}
\alias{changeWell}
\alias{changeWell,Well,Well-method}
\title{Change a well position}
\usage{
changeWell(from, to)

\S4method{changeWell}{Well,Well}(from, to)
}
\arguments{
\item{from}{A \code{\link{Well}} object.}

\item{to}{A \code{Well} object.}
}
\value{
Returns a \code{Well} object.
}
\description{
Makes sure that the updated well position is in the same plate format as the
original one.
}
\examples{

changeWell( Well("A01")
          , Well("B03", plateFormat = "undefined"))

}
