% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\name{Column}
\alias{Column}
\title{The column of a well}
\usage{
Column(object)
}
\arguments{
\item{object}{The [`Well`] object from which column coordinates are extracted.}
}
\value{
A column name in character format
}
\description{
Given a well object, tells in which column the well is.
}
\examples{
well <- Well("A01")
Column(well)

}
