% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispensing.R
\name{dispense_pairs}
\alias{dispense_pairs}
\title{Dispense pair combinations}
\usage{
dispense_pairs(
  plate,
  set1,
  set2 = set1,
  volume,
  n_per_pair = 1L,
  start = "A01"
)
}
\arguments{
\item{plate}{A \code{\link{Plate}} object.}

\item{set1}{First set of reagents.}

\item{set2}{Second set of reagents.}

\item{volume}{Fixed volume to dispense.}

\item{n_per_pair}{Number of replicates per pair.}

\item{start}{Starting well.}
}
\value{
Returns a \code{Plate} object with columns documenting the reagents to
be dispensed.
}
\description{
Functions to dispense combinations of all the pairs of two sets of reagents,
for instance forward and reverse PCR primers.
}
\details{
The names of the reagents must match the names that will be used in the
source plate.
}
\examples{
examplePlate |>
  dispense_pairs(c("F01", "F02", "F03"), c("R01", "R02"), volume = 20) |> head(6)
examplePlate |>
  dispense_pairs(c("F01", "F02"), volume = 10) |> head(6)
examplePlate |>
  dispense_pairs("F01", c("R01", "R02"), volume = 10, n = 3) |> head(6)

}
\seealso{
\code{\link[=titration_curve]{titration_curve()}}

Other Dispensing functions: 
\code{\link{set_block}()}
}
\concept{Dispensing functions}
