% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo.R
\docType{class}
\name{Echo-class}
\alias{Echo-class}
\alias{Echo}
\alias{as.character,Echo-method}
\alias{show,Echo-method}
\title{Echo state}
\usage{
\S4method{as.character}{Echo}(x)

\S4method{show}{Echo}(object)
}
\arguments{
\item{x}{An \code{\link{Echo}} object.}

\item{object}{An \code{\link{Echo}} object.}
}
\description{
Objects of this class represent the state of an Echo machine loaded with
a source and a destination plate, and with its transducer at a given
position.
}
\details{
The model of the echo machine determines the resolution of the liquid
transfers: 2.5 nL increments for the \dQuote{555} and 25 nL increments for
the \dQuote{525}.
}
\section{Slots}{

\describe{
\item{\code{source}}{A source \code{\link{Plate}}.}

\item{\code{destination}}{A destination \code{\link{Plate}}.}

\item{\code{transducer}}{A \code{\link{Transducer}} state.}

\item{\code{model}}{The model of the Echo machine.}

\item{\code{log}}{A log of the transfers}
}}

\examples{
plate <- Plate( type       =   "384"
              , deadVolume =  10000
              , maxVolume  = 100000)

transducer <- Transducer( source      = Well("A01")
                        , destination = Well("A01", plateFormat = "96"))

echo <- Echo( source      = plate
            , destination = plate
            , transducer  = transducer
            , model       = "525")
echo

}
