% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplePlate}
\alias{examplePlate}
\title{Example \code{Plate} object}
\format{
An object of class \code{Plate} with 384 rows and 2 columns.
}
\usage{
examplePlate
}
\description{
This object represents a 384-well plate with a dead volume of 10 microliters
and a maximum volume of 100 microliters.  It is filled with reagents in its
three first wells.
}
\keyword{datasets}
