% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleTransducer}
\alias{exampleTransducer}
\title{Example \code{Transducer} object}
\format{
An object of class \code{Transducer} of length 1.
}
\usage{
exampleTransducer
}
\description{
This \code{\link{Transducer}} object is about to transfer from well \code{A01} of a source
plate in 384-well format to well \code{A01} of a destination plate in 96-well
format.
}
\seealso{
Other example data: 
\code{\link{exampleEcho}},
\code{\link{examplePlate}}
}
\concept{example data}
\keyword{datasets}
