% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCR.R
\name{load_qPCR_results}
\alias{load_qPCR_results}
\alias{load_qPCR_meltcurve}
\alias{load_qPCR_raw}
\title{Load CT values from qPCR results file}
\usage{
load_qPCR_results(
  file,
  expName = NA,
  machine = c(NA, "VIA", "QS5"),
  format = "Excel",
  skip = 43,
  n_max = 384
)

load_qPCR_meltcurve(
  file,
  expName = NA,
  machine = c(NA, "VIA", "QS5"),
  format = "Excel",
  skip = 42
)

load_qPCR_raw(
  file,
  expName = NA,
  machine = c(NA, "VIA", "QS5"),
  format = "Excel",
  skip = 42,
  cycles = 40
)
}
\arguments{
\item{file}{Path to the file to open.}

\item{expName}{The name of the experiment.}

\item{machine}{The type of the machine that produced the results file.}

\item{format}{The format of the file.  At the moment only \emph{Excel} is supported.}

\item{skip}{The number of lines to skip before reading the data.}

\item{n_max}{The number of lines to read.}

\item{cycles}{The number of PCR cycles.}
}
\value{
Returns a \code{\link[tibble:tibble]{tibble::tibble()}} object reporting well coordinates, CT
value and melting temperature as computed by the machine, and optionally
the name of the experiment.
}
\description{
Loads CT values from the results file produced by a real-time quantitative
PCR (RT-qPCR) machine.
}
\note{
This function needs the \emph{readxl} package.
}
\seealso{
Other qPCR functions: 
\code{\link{calc_qPCR_TM}()},
\code{\link{calc_qPCR_mod}()}

Other qPCR functions: 
\code{\link{calc_qPCR_TM}()},
\code{\link{calc_qPCR_mod}()}

Other qPCR functions: 
\code{\link{calc_qPCR_TM}()},
\code{\link{calc_qPCR_mod}()}
}
\author{
Charles Plessy
}
\concept{qPCR functions}
