% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo.R
\name{moveTransducer}
\alias{moveTransducer}
\alias{moveTransducer,Transducer,Well,missing-method}
\alias{moveTransducer,Transducer,missing,Well-method}
\alias{moveTransducer,Transducer,Well,Well-method}
\alias{moveTransducer,Echo,Well,missing-method}
\alias{moveTransducer,Echo,missing,Well-method}
\alias{moveTransducer,Echo,Well,Well-method}
\title{Change the state of the transducer}
\usage{
moveTransducer(transducer, source, destination)

\S4method{moveTransducer}{Transducer,Well,missing}(transducer, source, destination)

\S4method{moveTransducer}{Transducer,missing,Well}(transducer, source, destination)

\S4method{moveTransducer}{Transducer,Well,Well}(transducer, source, destination)

\S4method{moveTransducer}{Echo,Well,missing}(transducer, source, destination)

\S4method{moveTransducer}{Echo,missing,Well}(transducer, source, destination)

\S4method{moveTransducer}{Echo,Well,Well}(transducer, source, destination)
}
\arguments{
\item{transducer}{A \code{\link{Transducer}} object.}

\item{source}{A \code{\link{Well}} object.}

\item{destination}{A \code{Well} object.}
}
\value{
Returns a transducer object in the new state.
}
\description{
Moves the tranducer to a new source position, or moves the destination plate,
or both
}
\examples{
moveTransducer(exampleTransducer, Well("A02"))
moveTransducer(exampleTransducer, destination = Well("A02", plateFormat = "96"))
moveTransducer( exampleTransducer
              , source      = Well("A02")
              , destination = Well("A02", plateFormat = "96"))

moveTransducer(exampleEcho, source = Well("A02"))

moveTransducer(exampleEcho, destination = Well("A02", plateFormat = "96"))

moveTransducer( exampleEcho
              , source      = Well("A02")
              , destination = Well("A02", plateFormat = "96"))

}
\seealso{
Other Dispenser functions: 
\code{\link{Echo-class}},
\code{\link{Transducer-class}},
\code{\link{changeWell}()},
\code{\link{planTransfers}()},
\code{\link{showLogs}()},
\code{\link{transfer}()}
}
\concept{Dispenser functions}
