% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\docType{data}
\name{N}
\alias{N}
\title{Avogadro's constant}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
N
}
\description{
Number of molecules in a mole.  Not exported, but made available through
the \code{:::} operator for convenience.
}
\examples{
topcorneR:::N
prettyNum(topcorneR:::N, digits = 9)
}
\seealso{
\url{https://en.wikipedia.org/wiki/Avogadro_constant}

Other Accessory functions: 
\code{\link{titration_curve}()}
}
\concept{Accessory functions}
\keyword{datasets}
