% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\name{nextWell}
\alias{nextWell}
\title{Move to the next well}
\usage{
nextWell(well)
}
\arguments{
\item{well}{A \code{\link{Well}} object.}
}
\value{
A \code{\link{Well}} object.
}
\description{
From left to right, and up to down, give the position of the next well.
}
\examples{

nextWell(Well("A12", plateFormat = "96"))
nextWell(Well("A12", plateFormat = "384"))
nextWell(Well("A24", plateFormat = "384"))

}
