% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo.R
\name{planTransfers}
\alias{planTransfers}
\title{Plan Echo transfers}
\usage{
planTransfers(source, destination, plan, model = "525")
}
\arguments{
\item{source}{A source \code{\link{Plate}}.}

\item{destination}{A destination \code{Plate}.}

\item{plan}{A \code{Plate} representing the additions to the
destination plate.}

\item{model}{Which Echo model to use.}
}
\value{
A transfer log
}
\description{
Given a source plate, a target plate and the representation of a filled
target plate, calculate the liquid transfers necessary perform the filling.
}
\details{
Each reagent of the plan plate is transferred one after the other.  Once
a reagent is selected, a source well is sought for the transfer and the
transducer is moved to it.
}
\examples{
sourcePlate <- examplePlate
head(sourcePlate)

planPlate <- Plate("384") |>
  setWell(Well("A01"), "dNTP",   50) |>
  setWell(Well("A02"), "dNTP",  100) |>
  setWell(Well("A01"), "buffer", 50)
head(planPlate)

destinationPlate <- Plate("384", deadVolume = 1e4, maxVolume = 1e5)

echo <- planTransfers(sourcePlate, destinationPlate, planPlate)
echo
echo |> showLogs()

}
\seealso{
\code{\link{showLogs}}
}
