% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateMap.R
\name{plateMap}
\alias{plateMap}
\alias{plateMap_all}
\title{Plot the contents of a plate}
\usage{
plateMap(plate, x, title = x)

plateMap_all(plate)
}
\arguments{
\item{plate}{a \code{\link{Plate}} object.}

\item{x}{One reagent of a plate.}

\item{title}{a title for the plot.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot}} object.
}
\description{
Plot the contents of a plate
}
\seealso{
Other Plate functions: 
\code{\link{Plate-class}},
\code{\link{PlateTypeToWellNames}()},
\code{\link{as.character,Well-method}},
\code{\link{nextWell}()},
\code{\link{plateWellVolume}()},
\code{\link{randomise}()},
\code{\link{seekReagent}()},
\code{\link{setWell}()},
\code{\link{set_block}()},
\code{\link{sourceReagent}()}
}
\author{
Charles Plessy
}
\concept{Plate functions}
