% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\name{plateWellVolume}
\alias{plateWellVolume}
\title{Get reagent volume}
\usage{
plateWellVolume(plate, well, what)
}
\arguments{
\item{plate}{A \code{\link{Plate}} object.}

\item{well}{A \code{\link{Well}} object.}
}
\description{
In a given plate, get the volume of the reagent contained in a given well.
}
\examples{

destPlate <- Plate(plate = tibble::tibble(well = platetools::num_to_well(1:384, plate = "384")))
destPlate \%<>\%
  setWell(Well(well = "A01", plateFormat = "384"), "dNTP", 50) \%>\%
  setWell(Well(well = "A02", plateFormat = "384"), "dNTP", 100) \%>\%
  setWell(Well(well = "A01", plateFormat = "384"), "buffer", 50)

destPlate \%>\% plateWellVolume(Well(well = "A01"))
destPlate \%>\% plateWellVolume(Well(well = "A01"), "dNTP")
destPlate \%>\% plateWellVolume(Well(well = "A01"), "buffer")

destPlate \%>\% plateWellVolume(Well(well = "A02"))
destPlate \%>\% plateWellVolume(Well(well = "A02"), "dNTP")
destPlate \%>\% plateWellVolume(Well(well = "A02"), "buffer")

}
