% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\name{randomise}
\alias{randomise}
\alias{randomise,Plate,missing,missing-method}
\alias{randomise,Plate,numeric,missing-method}
\alias{randomise,Plate,missing,numeric-method}
\title{Randomise reaction coordinates}
\usage{
randomise(plate, seed, vector)

\S4method{randomise}{Plate,missing,missing}(plate, seed, vector)

\S4method{randomise}{Plate,numeric,missing}(plate, seed, vector)

\S4method{randomise}{Plate,missing,numeric}(plate, seed, vector)
}
\arguments{
\item{plate}{A \code{\link{Plate}} object.}

\item{seed}{An integer number to set the seed (optional).}

\item{vector}{A (usually random) vector to sort the plate with (optional).}
}
\value{
a \code{Plate} object in which the randomisation vector was stored in
its \code{metadata} slot.
}
\description{
Randomise the coordinates of the reactions in a plate, so that reactions with
similar parameters are less likely to be affected by the same spatial
technical biases.  This is also useful when performing plate replicates.
Finally, the coordinates of negative controls will usually make a
non-symmetric pattern that helps to detect when a plate was rotated
accidentally.
}
\note{
Working with random seeds is tricky.  In particular, pay attention that
if you set the random seed just before running the \code{randomise} function, its
results will become deterministic.
}
\examples{
# Toy example of a 6-well plate containing a gradient increase of a reagent
p <- Plate(type = "6") |> set_block("A01~B03", "reagent", 1:6 * 10)
head(p)

# Setting the same seed returns the same randomisation
randomise(p, seed = 1022) |> head()
randomise(p, seed = 1022) |> head()

# Not providing a seed returns a really random order
randomise(p) |> head()
randomise(p) |> head()

# The random order can also be passed with the vector argument
randomise(p, vector = sample(1:6)) |> head()

# This means also that order can also be forced
randomise(p, vector = 6:1) |> head()

# The randomisation vector is stored in the Plate object's metadata
p <- randomise(p)
p@metadata$randomisation_vector

}
\seealso{
Other Plate functions: 
\code{\link{PlateTypeToWellNames}()},
\code{\link{as.character,Well-method}},
\code{\link{nextWell}()},
\code{\link{plateWellVolume}()},
\code{\link{seekReagent}()},
\code{\link{setWell}()},
\code{\link{set_block}()},
\code{\link{sourceReagent}()}
}
\author{
Charles Plessy
}
\concept{Plate functions}
