% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\name{Row}
\alias{Row}
\title{The row of a well}
\usage{
Row(object)
}
\value{
A row name in character format.
}
\description{
Given a well object, tells in which row the well is.
}
\examples{
well <- Well(well = "A01", plateFormat = "384")
Row(well)

}
