% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R, R/echo.R
\name{seekReagent}
\alias{seekReagent}
\alias{seekReagent,Plate,character,Well-method}
\alias{seekReagent,Plate,character,missing-method}
\alias{seekReagent,Echo,character,Well-method}
\alias{seekReagent,Echo,character,missing-method}
\title{Find a well that can provide enough reagent}
\usage{
seekReagent(object, reagent, start)

\S4method{seekReagent}{Plate,character,Well}(object, reagent, start)

\S4method{seekReagent}{Plate,character,missing}(object, reagent, start)

\S4method{seekReagent}{Echo,character,Well}(object, reagent, start)

\S4method{seekReagent}{Echo,character,missing}(object, reagent, start)
}
\arguments{
\item{object}{A \code{\link{Plate}} object.}

\item{reagent}{A reagent name.}

\item{start}{A \code{\link{Well}} object (to avoid backtracking).}
}
\value{
a \code{Well} object.
}
\description{
Given a \code{\link{Plate}} object, check if the reagent is available, and
if yes, in which well.
}
\examples{
sourcePlate <- Plate( type       = "384"
                    , deadVolume = 10000
                    , maxVolume  = 100000)

sourcePlate <- sourcePlate |>
  setWell(Well("A01"), "dNTP",   100000) |>
  setWell(Well("A02"), "dNTP",   100000) |>
  setWell(Well("A03"), "buffer", 100000)

seekReagent(sourcePlate, "buffer")

}
