% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\name{setWell}
\alias{setWell}
\title{Set the contents of a well}
\usage{
setWell(plate, well, what, volume)
}
\arguments{
\item{plate}{A [`Plate`] object.}

\item{well}{A [`Well`] object.}

\item{what}{A reagent name (character value).}

\item{volume}{A volume in nanoliters (numeric value).}
}
\value{
Returns a [`Plate`] object.
}
\description{
Set the contents of a well
}
\examples{
example("Plate-class")
sourcePlate <- plate
A01 <- Well(well = "A01", plateFormat = "384")
sourcePlate <- setWell(sourcePlate, A01, "dNTP", 100000)
sourcePlate <- setWell(sourcePlate, Well(well = "A02", plateFormat = "384"), "dNTP", 100000)
sourcePlate <- setWell(sourcePlate, Well(well = "A03", plateFormat = "384"), "buffer", 100000)
DataFrame(sourcePlate)
}
