% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo.R
\name{showLogs}
\alias{showLogs}
\alias{showLogs,Echo-method}
\title{Show transfer logs}
\usage{
showLogs(echo)

\S4method{showLogs}{Echo}(echo)
}
\arguments{
\item{echo}{An \code{Echo} object.}
}
\description{
Extract the logs from an \code{Echo} object and formats them as a
table.
}
\examples{

showLogs(exampleEcho)

}
