% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\name{sourceReagent}
\alias{sourceReagent}
\title{Get reagent name}
\usage{
sourceReagent(plate, well)
}
\arguments{
\item{plate}{A [`Plate`] object.}

\item{well}{A [`Well`] object.}
}
\description{
In a source plate, get the name of the reagent contained in a given well.
}
\examples{

sourcePlate <- Plate(wells = DataFrame(well = platetools::num_to_well(1:384, plate = 384)))
sourcePlate \%<>\%
  setWell(Well(well = "A01", plateFormat = "384"), "dNTP", 100000) \%>\%
  setWell(Well(well = "A02", plateFormat = "384"), "dNTP", 100000) \%>\%
  setWell(Well(well = "A03", plateFormat = "384"), "buffer", 100000)

sourcePlate \%>\% sourceReagent(Well(well = "A01"))
sourcePlate \%>\% sourceReagent(Well(well = "A03"))
sourcePlate \%>\% sourceReagent()

}
\concept{Plate method}
