% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{titration_curve}
\alias{titration_curve}
\title{Titration curve}
\usage{
titration_curve(steps = 12, maxvol = 1000, dropvolume = 2.5, zero = TRUE)
}
\arguments{
\item{steps}{The number of points in the titration curve.}

\item{maxvol}{The maximum volume to reach.}

\item{dropvolume}{The volume of a droplet, in nanoliters.}

\item{zero}{Start with a negative control with a volume of zero.}
}
\value{
A numeric vector of length \code{steps} indicating the volumes to transfer
in order to cover a volume range in a logarithmic manner.
}
\description{
Given a number of \emph{steps} and a target \emph{maximal volume}, this function will
return an increasing serie of volumes that cover the range in an
approximately logarithmic progression.  Each volume can be dispensed, as they
are guaranteed to be multiple of the \emph{droplet's volume}.  The range can be
forced to start at zero if a negative control is needed.
}
\examples{
titration_curve(12)
titration_curve(12, zero = FALSE)
titration_curve(12, dropvolume = 25)
titration_curve(12, dropvolume = 25, maxvol = 300)

}
\seealso{
Other Accessory functions: 
\code{\link{N}}
}
\author{
Charles Plessy
}
\concept{Accessory functions}
