% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo.R
\docType{class}
\name{Transducer-class}
\alias{Transducer-class}
\alias{Transducer}
\alias{as.character,Transducer-method}
\alias{show,Transducer-method}
\title{Transducer state}
\usage{
\S4method{as.character}{Transducer}(x)

\S4method{show}{Transducer}(object)
}
\arguments{
\item{x}{A \code{\link{Transducer}} object}

\item{object}{An \code{\link{Transducer}} object.}
}
\description{
Objects of this class represent the state of the transduction system in the
Echo machine: the position of the transducer in the coordinate system of the
source plate, and the position of the target plate on top of the transducer.
}
\section{Slots}{

\describe{
\item{\code{source}}{A \code{\link{Well}} in the source plate.}

\item{\code{destination}}{A \code{\link{Well}} in the destination plate.}
}}

\examples{
Transducer(
  source      = Well("A01"),
  destination = Well("A01", plateFormat = "96"))

}
