% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo.R
\name{transfer}
\alias{transfer}
\alias{transfer,Echo,numeric-method}
\title{Transfer volume from a source to a destination plate}
\usage{
transfer(echo, volume)

\S4method{transfer}{Echo,numeric}(echo, volume)
}
\arguments{
\item{echo}{An \code{\link{Echo}} object.}

\item{volume}{Volume to transfer.

The \code{what} parameter must correspond to a column name in the source
plate.}
}
\value{
Returns An \code{\link{Echo}} object with the \code{source}, \code{destination} and \code{log}
slots updated.
}
\description{
Reduce the volume of a reagent in the source plate at the coordinates of
the transducer, add that volume to the destination plate,
and register the transfer in the log.
}
\examples{

sourcePlate <- examplePlate
DataFrame(sourcePlate) |> head()

destinationPlate <- Plate("96", deadVolume = 1e4, maxVolume = 1e5)

transducer <- Transducer( source      = Well("A01")
                        , destination = Well("A01", plateFormat = "96"))

echo <- Echo( source      = sourcePlate
            , destination = destinationPlate
            , transducer  = transducer
            , model       = "525")

echo |> showLogs()

echo <- echo |>
  transfer(10) |>
  transfer(10) |>
  moveTransducer(destination = Well("A02", plateFormat = "96")) |>
  transfer(5)

echo |> showLogs()

}
