% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\name{validRow}
\alias{validRow}
\alias{validColumn}
\title{Tests the well's row name is valid}
\usage{
validRow(w)

validColumn(w)
}
\arguments{
\item{w}{A \code{\link{Well}}.}
}
\value{
TRUE if valid, otherwise returns an error

TRUE if valid, otherwise returns an error
}
\description{
Tests the well's row name is valid

Tests the well's column number is valid
}
\seealso{
Other Well functions: 
\code{\link{PlateTypeToWellNames}()},
\code{\link{Well-class}},
\code{\link{as.character,Well-method}},
\code{\link{nextWell}()}
}
\concept{Well functions}
