% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\name{validColumn}
\alias{validColumn}
\title{Tests the well's column number is valid}
\usage{
validColumn(w)
}
\value{
TRUE if valid, otherwise returns an error
}
\description{
Tests the well's column number is valid
}
\examples{
validColumn(Well("A01"))
}
