% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\docType{class}
\name{Well-class}
\alias{Well-class}
\alias{Well}
\title{Well coordinates}
\description{
Objects of this class represent a position in a microwell plate.  This class
exists so that the validity of the coordinates can be tested (for instance
\code{M16} will be rejected for a 96-well plate).
}
\section{Slots}{

\describe{
\item{\code{well}}{A position in the microwell plate.}

\item{\code{plateFormat}}{A plate format.}
}}

\examples{
well <- Well("A01", plateFormat = "384")
well
as.character(well)
show(well)
Row(well)
Column(well)

}
\seealso{
Other Well functions: 
\code{\link{PlateTypeToWellNames}()},
\code{\link{as.character,Well-method}},
\code{\link{nextWell}()},
\code{\link{validRow}()}
}
\concept{Well functions}
