% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleEcho}
\alias{exampleEcho}
\title{Example \code{Echo} object}
\format{
An object of class \code{Echo} of length 1.
}
\usage{
exampleEcho
}
\description{
This \code{\link{Echo}} object represents an Echo 525 machine loaded with a 384-well
source plate from the \code{\link{examplePlate}} object and a 96-well destination
plate, with its transducer ready to transfer between from source well \code{A01}
to destination well \code{A01}.
}
\seealso{
Other example data: 
\code{\link{examplePlate}},
\code{\link{exampleTransducer}}
}
\concept{example data}
\keyword{datasets}
