% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\docType{class}
\name{Plate-class}
\alias{Plate-class}
\alias{.Plate}
\alias{Plate}
\alias{head,Plate-method}
\alias{as.character,Plate-method}
\alias{show,Plate-method}
\title{A multiwell plate}
\usage{
\S4method{head}{Plate}(x, n = 6L, ...)

\S4method{as.character}{Plate}(x)

\S4method{show}{Plate}(object)
}
\arguments{
\item{x}{A \code{\link{Plate}} object.}

\item{n}{The number of wells to be displayed.}

\item{...}{arguments to be passed to or from other methods.}

\item{object}{A \code{\link{Plate}} object}
}
\description{
Objects of this class represent a multiwell plate.
}
\section{Slots}{

\describe{
\item{\code{type}}{A plate type such as \code{"384"}, \code{"96"}, etc.`}

\item{\code{deadVolume}}{The dead volume of wells in this plate.}

\item{\code{maxVolume}}{The maximum volume of wells in this plate.}
}}

\examples{
plate <- Plate( type       = "384"
              , deadVolume = 10000
              , maxVolume  = 100000)
plate

}
\seealso{
Other Plate functions: 
\code{\link{PlateTypeToWellNames}()},
\code{\link{as.character,Well-method}},
\code{\link{nextWell}()},
\code{\link{plateMap}()},
\code{\link{plateWellVolume}()},
\code{\link{randomise}()},
\code{\link{seekReagent}()},
\code{\link{setWell}()},
\code{\link{set_block}()},
\code{\link{sourceReagent}()}
}
\concept{Plate functions}
