% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\docType{class}
\name{Plate-class}
\alias{Plate-class}
\alias{.Plate}
\alias{Plate}
\title{A multiwell plate}
\description{
Objects of this class represent a multiwell plate.
}
\section{Slots}{

\describe{
\item{\code{type}}{A plate type such as `"384"`, `"96"`, etc.`}

\item{\code{deadVolume}}{The dead volume of wells in this plate.}

\item{\code{maxVolume}}{The maximum volume of wells in this plate.}
}}

\examples{
plate <- Plate( type       = "384"
              , deadVolume = 10000
              , maxVolume  = 100000)
plate

}
