% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateMap.R
\name{plateMap}
\alias{plateMap}
\alias{plateMap_all}
\title{Plot the contents of a plate}
\usage{
plateMap(plate, x, title = x)

plateMap_all(plate)
}
\arguments{
\item{plate}{a \code{\link{Plate}} object.}

\item{x}{One reagent of a plate.}

\item{title}{a title for the plot.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot}} object.
}
\description{
Plot the contents of a plate
}
\author{
Charles Plessy
}
