% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\name{PlateTypeToWellNames}
\alias{PlateTypeToWellNames}
\title{Well names of plate type}
\usage{
PlateTypeToWellNames(type = c("6", "96", "384"))
}
\arguments{
\item{type}{The name a plate type (character)}
}
\value{
Returns a character vector of well names.
}
\description{
Lists the well names of a known well plate format, from left to right
and top to bottom.
}
\examples{
PlateTypeToWellNames("6")
PlateTypeToWellNames("96")
PlateTypeToWellNames("384")

}
\seealso{
Other Plate functions: 
\code{\link{Plate-class}},
\code{\link{as.character,Well-method}},
\code{\link{nextWell}()},
\code{\link{plateMap}()},
\code{\link{plateWellVolume}()},
\code{\link{randomise}()},
\code{\link{seekReagent}()},
\code{\link{setWell}()},
\code{\link{set_block}()},
\code{\link{sourceReagent}()}

Other Well functions: 
\code{\link{Well-class}},
\code{\link{as.character,Well-method}},
\code{\link{nextWell}()},
\code{\link{validRow}()}
}
\concept{Plate functions}
\concept{Well functions}
