% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\name{plateWellVolume}
\alias{plateWellVolume}
\title{Get reagent volume}
\usage{
plateWellVolume(plate, well, what)
}
\arguments{
\item{plate}{A [`Plate`] object.}

\item{well}{A [`Well`] object.}

\item{what}{The name of a reagent.}
}
\description{
In a given plate, get the volume of the reagent contained in a given well.
}
\examples{

destPlate <- Plate( type       = "384"
                  , deadVolume = 10000
                  , maxVolume  = 100000)
destPlate <- destPlate |>
  setWell(Well("A01"), "dNTP",   50) |>
  setWell(Well("A02"), "dNTP",  100) |>
  setWell(Well("A01"), "buffer", 50)

destPlate |> plateWellVolume(Well("A01"))
destPlate |> plateWellVolume(Well("A01"), "dNTP")
destPlate |> plateWellVolume(Well("A01"), "buffer")

destPlate |> plateWellVolume(Well("A02"))
destPlate |> plateWellVolume(Well("A02"), "dNTP")
destPlate |> plateWellVolume(Well("A02"), "buffer")

}
