% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\name{seekReagent}
\alias{seekReagent}
\title{Find a well that can provide enough reagent}
\usage{
seekReagent(object, reagent, start)
}
\arguments{
\item{object}{A \code{\link{Plate}} object.}

\item{reagent}{A reagent name.}

\item{start}{A \code{\link{Well}} object (to avoid backtracking).}
}
\value{
a \code{\link{Well}} object.
}
\description{
Given a \code{\link{Plate}} object, check if the reagent is available, and
if yes, in which well.
}
\examples{
sourcePlate <- Plate(plate = tibble::tibble(well = platetools::num_to_well(1:384, plate = "384")))
sourcePlate \%<>\%
  setWell(Well(well = "A01", plateFormat = "384"), "dNTP", 100000) \%>\%
  setWell(Well(well = "A02", plateFormat = "384"), "dNTP", 100000) \%>\%
  setWell(Well(well = "A03", plateFormat = "384"), "buffer", 100000)

seekReagent(sourcePlate, "buffer")

}
