% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\name{seekReagent}
\alias{seekReagent}
\title{Find a well that can provide enough reagent}
\usage{
seekReagent(object, reagent, start)
}
\arguments{
\item{object}{A [`Plate`] object.}

\item{reagent}{A reagent name.}

\item{start}{A [`Well`] object (to avoid backtracking).}
}
\value{
a `Well` object.
}
\description{
Given a [`Plate`] object, check if the reagent is available, and
if yes, in which well.
}
\examples{
sourcePlate <- Plate( type       = "384"
                    , deadVolume = 10000
                    , maxVolume  = 100000)

sourcePlate <- sourcePlate |>
  setWell(Well("A01"), "dNTP",   100000) |>
  setWell(Well("A02"), "dNTP",   100000) |>
  setWell(Well("A03"), "buffer", 100000)

seekReagent(sourcePlate, "buffer")

}
