% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R, R/echo.R
\name{seekReagent}
\alias{seekReagent}
\alias{seekReagent,Plate,character,Well-method}
\alias{seekReagent,Plate,character,missing-method}
\alias{seekReagent,Echo,character,Well-method}
\alias{seekReagent,Echo,character,missing-method}
\title{Find a well that can provide enough reagent}
\usage{
seekReagent(object, reagent, start)

\S4method{seekReagent}{Plate,character,Well}(object, reagent, start)

\S4method{seekReagent}{Plate,character,missing}(object, reagent, start)

\S4method{seekReagent}{Echo,character,Well}(object, reagent, start)

\S4method{seekReagent}{Echo,character,missing}(object, reagent, start)
}
\arguments{
\item{object}{A \code{\link{Plate}} object.}

\item{reagent}{A reagent name.}

\item{start}{A \code{\link{Well}} object (to avoid backtracking).}
}
\value{
a \code{Well} object.
}
\description{
Given a \code{\link{Plate}} object, check if the reagent is available, and
if yes, in which well.  The search can be restricted to exclude the wells
positioned before (in alphanumeric order) an arbitrary position, to avoid
backtracking to wells already emptired.
}
\examples{
seekReagent(examplePlate, "buffer")

}
\seealso{
Other Plate functions: 
\code{\link{PlateTypeToWellNames}()},
\code{\link{as.character,Well-method}},
\code{\link{nextWell}()},
\code{\link{plateWellVolume}()},
\code{\link{setWell}()},
\code{\link{set_block}()},
\code{\link{sourceReagent}()}
}
\concept{Plate functions}
