% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\name{setWell}
\alias{setWell}
\alias{setWell,Plate,Well,character,numeric-method}
\alias{setWell,Plate,Well,character,logical-method}
\title{Set the contents of a well}
\usage{
setWell(plate, well, what, volume)

\S4method{setWell}{Plate,Well,character,numeric}(plate, well, what, volume)

\S4method{setWell}{Plate,Well,character,logical}(plate, well, what, volume)
}
\arguments{
\item{plate}{A \code{\link{Plate}} object.}

\item{well}{A \code{\link{Well}} object.}

\item{what}{A reagent name (character value).}

\item{volume}{A volume in nanoliters (numeric value).}
}
\value{
Returns a \code{\link{Plate}} object.
}
\description{
Set the contents of a well
}
\examples{
example("Plate-class")
sourcePlate <- plate
A01 <- Well("A01")
sourcePlate <- setWell(sourcePlate, A01, "dNTP", 100000)
sourcePlate <- setWell(sourcePlate, Well(well = "A02", plateFormat = "384"), "dNTP", 100000)
sourcePlate <- setWell(sourcePlate, Well(well = "A03", plateFormat = "384"), "buffer", 100000)
DataFrame(sourcePlate)
}
