% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo.R
\name{showLogs}
\alias{showLogs}
\alias{showLogs,Echo-method}
\alias{exportLogs}
\alias{exportLogs,Echo-method}
\title{Show or export transfer logs}
\usage{
showLogs(echo)

\S4method{showLogs}{Echo}(echo)

exportLogs(echo)

\S4method{exportLogs}{Echo}(echo)
}
\arguments{
\item{echo}{An \code{Echo} object.}
}
\description{
Extract the logs from an \code{Echo} object and formats them as a
table.  The \code{exportLogs} function sets the column names expected by the
\emph{Echo Plate Reformat} software and discards the \code{what} column.
}
\examples{
showLogs(exampleEcho)
exportLogs(exampleEcho)

}
\seealso{
Other Dispenser functions: 
\code{\link{Echo-class}},
\code{\link{Transducer-class}},
\code{\link{changeWell}()},
\code{\link{moveTransducer}()},
\code{\link{planTransfers}()},
\code{\link{transfer}()}
}
\concept{Dispenser functions}
