% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\name{sourceReagent}
\alias{sourceReagent}
\title{Get reagent name}
\usage{
sourceReagent(plate, well)
}
\arguments{
\item{plate}{A [`Plate`] object.}

\item{well}{A [`Well`] object.}
}
\description{
In a source plate, get the name of the reagent contained in a given well.
}
\examples{

sourcePlate <- Plate( type       = "384"
                    , deadVolume = 10000
                    , maxVolume  = 100000)

sourcePlate <- sourcePlate |>
  setWell(Well("A01"), "dNTP",   100000) |>
  setWell(Well("A02"), "dNTP",   100000) |>
  setWell(Well("A03"), "buffer", 100000)

sourcePlate |> DataFrame() |> head()

sourcePlate |> sourceReagent(Well("A01"))
sourcePlate |> sourceReagent(Well("A03"))

}
\concept{Plate method}
