% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\name{sourceReagent}
\alias{sourceReagent}
\alias{sourceReagent,Plate,Well-method}
\alias{sourceReagent,Plate,missing-method}
\title{Get reagent name}
\usage{
sourceReagent(plate, well)

\S4method{sourceReagent}{Plate,Well}(plate, well)

\S4method{sourceReagent}{Plate,missing}(plate, well)
}
\arguments{
\item{plate}{A \code{\link{Plate}} object.}

\item{well}{A \code{\link{Well}} object.}
}
\description{
In a source plate, get the name of the reagent contained in a given well.
}
\examples{

sourcePlate <- Plate( type       = "384"
                    , deadVolume = 10000
                    , maxVolume  = 100000)

sourcePlate <- sourcePlate |>
  setWell(Well("A01"), "dNTP",   100000) |>
  setWell(Well("A02"), "dNTP",   100000) |>
  setWell(Well("A03"), "buffer", 100000)

sourcePlate |> DataFrame() |> head()

sourcePlate |> sourceReagent(Well("A01"))
sourcePlate |> sourceReagent(Well("A03"))
sourcePlate |> sourceReagent()

}
\seealso{
Other Plate functions: 
\code{\link{PlateTypeToWellNames}()},
\code{\link{as.character,Well-method}},
\code{\link{nextWell}()},
\code{\link{plateWellVolume}()},
\code{\link{seekReagent}()},
\code{\link{setWell}()}
}
\concept{Plate functions}
