% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\name{validRow}
\alias{validRow}
\alias{Column}
\title{Tests the well's row name is valid}
\usage{
validRow(w)

Column(object)
}
\value{
TRUE if valid, otherwise returns an error

A column name in character format
}
\description{
Given a well object, tells in which column the well is.
}
\examples{
validRow(Well("A01", "96"))
well <- Well("A01")
Column(well)

}
